skilltrack_pro/backend/requirements.txt
Flask>=2.0.0
supabase>=1.0.0
python-dotenv>=0.21.0
gunicorn
